🎅🏿 Add nawc volunteer role (#458)
* Add new role, remove intern role

* Add new file
