+++ client/src/components/RoleApproval.jsx
@@ -76,7 +76,7 @@ const RoleApproval = () => {
   const RoleMenu = (email) => (
     <Menu onClick={(e) => setNewRoleAndUser(e.key, email)}>
       <Menu.Item key="public">public</Menu.Item>
-      <Menu.Item key="intern">intern</Menu.Item>
+      <Menu.Item key="nawc volunteer">NAWC volunteer</Menu.Item>
       <Menu.Item key="admin">admin</Menu.Item>
     </Menu>
   );
